requirements.txt
wheel 
streamlit
python-dotenv
ipykernel
langchain
langchain-core
langchain-community
langchain-groq
langchain-huggingface
faiss-cpu
langchain-text-splitters
langsmith
sentence-transformers
huggingface-hub
torch
protobuf==3.20.*